Java Program that generate an isosceles triangle made of asterisks
Write a Java program that generates an isosceles triangle made of asterisks. The program should prompt the user to specify the size of the triangle. Make sure variable names are appropriate,and make sure there are comments for the weird variables like sideSize. Use good style.
Sample session:
Enter the size of the equal sides in an isosceles triangle: 6
*
**
***
****
*****
******